refactor /visualizations routes into separate files (#352)
* refactor /visualizations routes into separate files
* bug fixes
* pkg lock
* add a `GET /visualizations` route for pure REST purposes
