+++ src/routes/visualizations/{id}/styles.js
@@ -1,23 +1,13 @@
-const fs = require('fs-extra');
 const path = require('path');
+const fs = require('fs-extra');
 const Boom = require('@hapi/boom');
 const Joi = require('joi');
-const chartCore = require('@datawrapper/chart-core');
 const get = require('lodash/get');
+const chartCore = require('@datawrapper/chart-core');
+const { createFontEntries, compileCSS } = require('../../../publish/compile-css.js');
 const set = require('lodash/set');
-const { translate } = require('@datawrapper/service-utils/l10n');
 
-const { createFontEntries, compileCSS } = require('../publish/compile-css.js');
-
-module.exports = {
-    name: 'routes/visualizations',
-    version: '1.0.0',
-    register
-};
-
-async function register(server, options) {
-    server.app.scopes.add('visualization:read');
-
+module.exports = (server, options) => {
     const styleCache = server.cache({
         segment: 'vis-styles',
         expiresIn: 86400000 * 365 /* 1 year */,
@@ -26,34 +16,7 @@ async function register(server, options) {
 
     server.route({
         method: 'GET',
-        path: '/{id}',
-        options: {
-            auth: {
-                mode: 'try',
-                access: { scope: ['visualization:read'] }
-            }
-        },
-        handler: getVisualization
-    });
-
-    async function getVisualization(request, h) {
-        const { params, server, auth } = request;
-
-        const vis = server.app.visualizations.get(params.id);
-        if (!vis) return Boom.notFound();
-
-        // also include translated title
-        vis.__title = translate(vis.title, {
-            scope: vis.__plugin,
-            language: get(auth.artifacts, 'language') || 'en-US'
-        });
-
-        return h.response(vis);
-    }
-
-    server.route({
-        method: 'GET',
-        path: '/{id}/styles.css',
+        path: '/styles.css',
         options: {
             auth: {
                 mode: 'try'
@@ -70,7 +33,7 @@ async function register(server, options) {
 
     server.route({
         method: 'GET',
-        path: '/{id}/styles',
+        path: '/styles',
         options: {
             auth: {
                 mode: 'try'
@@ -148,40 +111,4 @@ async function register(server, options) {
             return { css, fonts };
         }
     }
-
-    server.route({
-        method: 'GET',
-        path: '/{id}/script.js',
-        options: {
-            auth: {
-                mode: 'try',
-                access: { scope: ['visualization:read'] }
-            }
-        },
-        handler: getVisualizationScript
-    });
-
-    async function getVisualizationScript(request, h) {
-        const { params, server } = request;
-
-        const { result, statusCode } = await server.inject({
-            url: `/v3/visualizations/${params.id}`,
-            validate: false
-        });
-
-        if (statusCode !== 200) {
-            return new Boom.Boom(result.message, result);
-        }
-
-        const file = result.script;
-        const { mtime } = await fs.stat(file);
-
-        /* https://hapi.dev/api/?v=19.1.1#-hentityoptions */
-        const response = h.entity({ modified: mtime });
-
-        if (response) return response;
-
-        const stream = fs.createReadStream(file, { encoding: 'utf-8' });
-        return h.response(stream).header('Content-Type', 'application/javascript');
-    }
-}
+};
